requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dotenv==0.9.9
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.27.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.0
six==1.17.0
smmap==5.0.2
streamlit==1.42.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
plotly==5.15.0
pgeocode==0.3.0
folium==0.14.0
streamlit-folium==0.10.0
numerize==0.12.0